getting started with the status
