+++ src/components/Home/Lifestyle-status/lifestyle.css
@@ -25,3 +25,174 @@
 	font-size: 11px;
 	color: var(--primary-color);
 }
+
+.lifestyle-modal {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    width: 70%;
+    border: 0;
+    border-radius: 1rem;
+    box-shadow: 0px 1px 2px 0px rgba(16, 24, 40, 0.05) !important;
+    background: white;
+    padding: 1rem 1.7rem;
+}
+
+.lifestyle-modal-header {
+    width: 100%;
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    align-items: center;
+    justify-content: space-between;
+    padding: 1rem;
+}
+
+.lifestyle-modal-logo {
+    display: flex;
+    align-items: center;
+    gap: .8rem;
+}
+
+.lifestyle-modal-logo img {
+    width: 40px;
+    height: 40px;
+}
+
+.lifestyle-modal-logo >h3 {
+    color: var(--primary-color);
+}
+
+.lifestyle-modal-cancel-btn {
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+}
+
+.lifestyle-main-display {
+    display: grid;
+    grid-template-columns: 240px 1fr;
+}
+
+.lifestyle-main-display >.left-display {
+    border-top:1px solid grey;
+}
+
+.lifestyle-main-display >.left-display button {
+    border: 1px solid grey;
+    border-radius: 6px;
+    display: flex;
+    align-items: center;
+    gap: .8rem;
+    margin-top: 1rem;
+    padding: 1rem;
+}
+
+.left-display button .btn-text-wrapper{
+    text-align: left;
+}
+.left-display >.friends-status {
+    max-height: 260px;
+    margin-top: .8rem;
+    overflow-y: scroll;
+}
+
+.status-content {
+    background: white;
+    padding: 2rem;
+    height: 200px;
+    width: 200px;
+}
+
+.status-logo {
+    position: absolute;
+    color: red;
+    top: 0px;
+    left: 6px;
+    bottom: 16px;
+    z-index: 10;
+    width: fit-content;
+    height: 48px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-weight: bold;
+}
+
+.lifestyle-swiper {
+    width: 100% !important;
+    height: 100%;
+    border: 1px solid red !important;
+}
+
+.lifestyle-swiper .swiper-slide-shadow-left,
+.lifestyle-swiper .swiper-slide-shadow-right
+{
+    border: 1px solid blue !important;
+    display: block;
+}
+
+.swiper-slide img {
+    display: block;
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+.lifestyle-status-next,
+.lifestyle-status-prev {
+        width: 40px;
+        height: 40px;
+        border-radius: 50%;
+        background-repeat: no-repeat;
+        background-size: 100% auto;
+        background-position: center;
+        background-color: white !important;
+}
+
+.swiper-button-next {
+    background-image: url('../../../assets/custom-icons/circle-arrow-right-solid.svg');
+}
+
+.swiper-button-prev {
+    background-image: url('../../../assets/custom-icons/circle-arrow-left-solid.svg');
+    background-repeat: no-repeat;
+}
+
+.swiper-pagination-bullet {
+    flex: 1;
+    height: 4px;
+    border-radius: 20px;
+    text-align: center;
+    line-height: 20px;
+    font-size: 8px;
+    color: #000;
+    opacity: 1;
+    background: linear-gradient(to right, var(--primary-color) 50%, #f1eeeece 50%);
+    background-size: 200% 100%;
+    background-position: right bottom;
+    transition: none;
+}
+
+.swiper-pagination-bullet-active {
+    transition: all 3s ease-in;
+    background-position: left bottom;
+}
+
+.swiper-pagination-bullets.swiper-pagination-horizontal {
+    top: 0px !important;
+    left: 94px !important;
+    background: rgba(0, 0, 0, 0.4) !important;
+    width: calc(100% - 94px) !important;
+    height: 40px !important;
+    display: flex !important;
+    flex-direction: row !important;
+    margin-top: 6px !important;
+    align-items: center !important;
+}
+
+@media screen and (max-width: 524px) {
+    .lifestyle-modal {
+        width: 80%;
+    }
+}
