+++ src/index.css
@@ -27,6 +27,8 @@ body {
   --background-main-color: rgba(251, 248, 255, 0.95);
   --primary: #d23369;
   --primary-color: #4f0da3;
+  --primary-color-light: #7f2bee;
+  --primary-color-dark: #220647;
   --primary-deep: #f06292;
 }
