+++ src/index.js
@@ -10,8 +10,8 @@ import "bootstrap/dist/css/bootstrap.min.css";
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <Router>
-      <App />
+        <Router>
+            <App />
     </Router>
   </React.StrictMode>
 );
