+++ src/pages/FeedDetail/feed-detail.css
@@ -2,6 +2,7 @@
     position: fixed;
     top: 88dvh;
     right: 30%;
+    border: 1px solid red;
 }
 
 @media screen and (max-width: 1150px){
